# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sales/{id}/confirm": {
      "patch": {
        "operationId": "confirmSale",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "sale ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm a sale",
        "tags": [
          "Sales"
        ]
      }
    },
    "/sales": {
      "get": {
        "operationId": "getAllSales",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of sales returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetSaleDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all sales",
        "tags": [
          "Sales"
        ]
      },
      "post": {
        "operationId": "createSale",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sale created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new sale",
        "tags": [
          "Sales"
        ]
      }
    },
    "/sales/available-products": {
      "get": {
        "operationId": "getAvailableProductsToSell",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of available products to sell returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAvailableProductsToSellDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get available products to sell",
        "tags": [
          "Sales"
        ]
      }
    },
    "/sales/{id}": {
      "get": {
        "operationId": "getOneSale",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Sale ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sale found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSaleDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Sale not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific sale",
        "tags": [
          "Sales"
        ]
      },
      "patch": {
        "operationId": "updateSale",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Sale ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSaleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sale updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a sale",
        "tags": [
          "Sales"
        ]
      },
      "delete": {
        "operationId": "deleteSale",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Sale ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sale deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a sale",
        "tags": [
          "Sales"
        ]
      }
    },
    "/sales/{id}/installments/mark-paid": {
      "patch": {
        "operationId": "markInstallmentPaid",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Sale ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkInstallmentPaidDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Installment marked as paid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark installment as paid",
        "tags": [
          "Sales"
        ]
      }
    },
    "/sales/{id}/status": {
      "patch": {
        "operationId": "updateSaleStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Sale ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSaleStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sale status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update sale status",
        "tags": [
          "Sales"
        ]
      }
    },
...
  },
  "info": {
    "title": "Luxis API",
    "description": "System API for Luxis",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GetSaleProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the sale product",
            "example": "123e4567-e89b-12d3-a456-426614174003"
          },
          "serialNumber": {
            "type": "string",
            "description": "The serial number of the product",
            "example": "SN123456789"
          },
          "salePrice": {
            "type": "string",
            "description": "The sale price of the product",
            "example": "500.00"
          },
          "modelName": {
            "type": "string",
            "description": "The name of the product model",
            "example": "iPhone 13 Pro Max"
          },
          "categoryName": {
            "type": "string",
            "description": "The name of the category",
            "example": "Smartphones"
          },
          "modelId": {
            "type": "string",
            "description": "The ID of the product model",
            "example": "123e4567-e89b-12d3-a456-426614174002"
          },
          "categoryId": {
            "type": "string",
            "description": "The ID of the category",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          }
        },
        "required": [
          "id",
          "serialNumber",
          "salePrice",
          "modelName",
          "categoryName",
          "modelId",
          "categoryId"
        ]
      },
      "GetSaleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the sale",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "customerId": {
            "type": "string",
            "description": "The ID of the customer",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "customerName": {
            "type": "string",
            "description": "The name of the customer",
            "example": "John Doe"
          },
          "customerPhone": {
            "type": "string",
            "description": "The phone number of the customer",
            "example": "+5511999999999"
          },
          "resellerId": {
            "type": "string",
            "description": "The ID of the reseller",
            "example": "123e4567-e89b-12d3-a456-426614174002"
          },
          "resellerName": {
            "type": "string",
            "description": "The name of the reseller",
            "example": "Reseller Name"
          },
          "products": {
            "description": "The products in the sale",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetSaleProductDto"
            }
          },
          "saleDate": {
            "format": "date-time",
            "type": "string",
            "description": "The date of the sale",
            "example": "2024-01-01"
          },
          "totalAmount": {
            "type": "string",
            "description": "The total amount of the sale",
            "example": "1000.00"
          },
          "paymentMethod": {
            "type": "string",
            "description": "The payment method of the sale",
            "enum": [
              "CASH",
              "PIX",
              "DEBIT",
              "CREDIT",
              "EXCHANGE"
            ],
            "example": "DEBIT"
          },
          "numberInstallments": {
            "type": "number",
            "description": "The number of installments",
            "example": 12
          },
          "status": {
            "type": "string",
            "description": "The status of the sale",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "CANCELLED",
              "INSTALLMENTS_PENDING",
              "INSTALLMENTS_PAID",
              "INSTALLMENTS_OVERDUE"
            ],
            "example": "INSTALLMENTS_PENDING"
          },
          "installmentsInterval": {
            "type": "number",
            "description": "The interval between installments in days",
            "example": 30
          },
          "installmentsPaid": {
            "type": "number",
            "description": "The number of installments paid",
            "example": 3
          }
        },
        "required": [
          "id",
          "customerId",
          "customerName",
          "customerPhone",
          "resellerId",
          "resellerName",
          "products",
          "saleDate",
          "totalAmount",
          "paymentMethod",
          "numberInstallments",
          "status",
          "installmentsInterval",
          "installmentsPaid"
        ]
      },
      "GetAvailableProductModelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the product model",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "modelName": {
            "type": "string",
            "description": "The name of the product model",
            "example": "Product Model Name"
          },
          "imageUrl": {
            "type": "string",
            "description": "The image URL of the product model",
            "example": "https://example.com/product-model.jpg"
          },
          "products": {
            "description": "The available products for the model",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetSaleProductDto"
            }
          }
        },
        "required": [
          "id",
          "modelName",
          "products"
        ]
      },
      "GetAvailableCategoryDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "description": "The ID of the category",
            "example": "a3bb189e-8bf9-3888-9912-ace4e6543002"
          },
          "categoryName": {
            "type": "string",
            "description": "The name of the category",
            "example": "Smartphones"
          },
          "models": {
            "description": "The available product models",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetAvailableProductModelDto"
            }
          }
        },
        "required": [
          "categoryId",
          "categoryName",
          "models"
        ]
      },
      "GetAvailableProductsToSellDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The available categories with their product models and products",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetAvailableCategoryDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "CASH",
          "PIX",
          "DEBIT",
          "CREDIT",
          "EXCHANGE"
        ],
        "description": "The payment method of the sale"
      },
      "CreateSaleDto": {
        "type": "object",
        "properties": {
          "productIds": {
            "description": "The IDs of the products",
            "example": [
              "123e4567-e89b-12d3-a456-426614174000",
              "123e4567-e89b-12d3-a456-426614174001"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "saleDate": {
            "format": "date-time",
            "type": "string",
            "description": "The date of the sale",
            "example": "2021-01-01"
          },
          "paymentMethod": {
            "description": "The payment method of the sale",
            "example": "CREDIT",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentMethod"
              }
            ]
          },
          "numberInstallments": {
            "type": "number",
            "description": "The number of installments of the sale",
            "default": 1
          },
          "installmentsInterval": {
            "type": "number",
            "description": "The interval of installments of the sale",
            "default": 0
          },
          "customerId": {
            "type": "string",
            "description": "The ID of the consumer",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "productIds",
          "saleDate",
          "paymentMethod",
          "numberInstallments",
          "installmentsInterval",
          "customerId"
        ]
      },
      "SaleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "customerId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "resellerId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174002"
          },
          "productIds": {
            "example": [
              "123e4567-e89b-12d3-a456-426614174003"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "saleDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "totalAmount": {
            "type": "string",
            "example": "1000.00"
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "CASH",
              "PIX",
              "DEBIT",
              "CREDIT",
              "EXCHANGE"
            ],
            "example": "DEBIT"
          },
          "numberInstallments": {
            "type": "number",
            "example": 12
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "CANCELLED",
              "INSTALLMENTS_PENDING",
              "INSTALLMENTS_PAID",
              "INSTALLMENTS_OVERDUE"
            ],
            "example": "CONFIRMED"
          },
          "installmentsInterval": {
            "type": "number",
            "example": 30
          },
          "installmentsPaid": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "id",
          "customerId",
          "resellerId",
          "productIds",
          "saleDate",
          "totalAmount",
          "paymentMethod",
          "numberInstallments",
          "status",
          "installmentsInterval",
          "installmentsPaid"
        ]
      },
      "UpdateSaleDto": {
        "type": "object",
        "properties": {
          "productIds": {
            "description": "The IDs of the products",
            "example": [
              "123e4567-e89b-12d3-a456-426614174000",
              "123e4567-e89b-12d3-a456-426614174001"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "productIds"
        ]
      },
      "MarkInstallmentPaidDto": {
        "type": "object",
        "properties": {
          "installmentNumber": {
            "type": "number",
            "description": "The number of installments to mark as paid",
            "example": 2
          }
        },
        "required": [
          "installmentNumber"
        ]
      },
      "UpdateSaleStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "The new status of the sale",
            "example": "INSTALLMENTS_PENDING",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "CANCELLED",
              "INSTALLMENTS_PENDING",
              "INSTALLMENTS_PAID",
              "INSTALLMENTS_OVERDUE"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}